src/execution/exec/ft_exec_path.c: norminette
